+++ mgh_model/src/sim.cpp
@@ -7,7 +7,7 @@ int main(int argc, char *argv[]) {
   // Initialize sim: parse params & make objects (MTs, kinesin, MAPs, etc)
   properties.wallace.InitializeSimulation(argv, &properties, &parameters);
   // Main KMC loop
-  for (int i_step{0}; i_step < parameters.n_steps; i_step++) {
+  for (unsigned long i_step{0}; i_step < parameters.n_steps; i_step++) {
     properties.wallace.UpdateTimestep(i_step);
     properties.kinesin4.RunKMC();
     properties.prc1.RunKMC();
